user registration logic init version
