+++ packages/client/src/components/SignUpPage/_utils/transformFormDataToDTO.ts
@@ -1,0 +1,14 @@
+import type { SignUpFormData, SignUpRequestDTO } from '../_types';
+
+export function transformFormDataToDTO(
+  formData: SignUpFormData
+): SignUpRequestDTO {
+  return {
+    first_name: formData.name,
+    second_name: formData.secondName,
+    login: formData.login,
+    email: formData.email,
+    password: formData.password,
+    phone: formData.phone,
+  };
+}
